Reorganized Shared.Storage folder
