+++ Content.Client/Storage/Visualizers/BagOpenCloseVisualizer.cs
@@ -1,6 +1,7 @@
 ﻿
 using Content.Shared.Stacks;
 using Content.Shared.Storage;
+using Content.Shared.Storage.Components;
 using JetBrains.Annotations;
 using Robust.Client.GameObjects;
 using Robust.Shared.GameObjects;
