+++ Content.Shared/Storage/Components/ItemMapperComponent.cs
@@ -3,12 +3,12 @@ using Robust.Shared.GameObjects;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Manager.Attributes;
 
-namespace Content.Shared.Storage.ItemCounter
+namespace Content.Shared.Storage.Components
 {
     [RegisterComponent]
-    public class ItemCounterComponent : Component, ISerializationHooks
+    public class ItemMapperComponent : Component, ISerializationHooks
     {
-        public override string Name => "ItemCounter";
+        public override string Name => "ItemMapper";
 
         [DataField("mapLayers")] public readonly Dictionary<string, SharedMapLayerData> MapLayers = new();
