+++ Content.Shared/Storage/Components/SharedBagOpenVisuals.cs
@@ -1,7 +1,7 @@
 using System;
 using Robust.Shared.Serialization;
 
-namespace Content.Shared.Storage
+namespace Content.Shared.Storage.Components
 {
     [Serializable, NetSerializable]
     public enum SharedBagOpenVisuals : byte
